Anime Fox/requirements.txt
requests==2.18.4
SQLAlchemy==1.2.2
beautifulsoup4==4.6.0
psycopg2-binary==2.7.4
PyYAML==3.12
